Automate loading of previous output file
Make get_halo() in function_NFW.py return the name of the file it writes to.
This is passed back to the next call of the function to load the previous data,
so manually setting loadprev isn't required any more, and we don't need to guess
the filename to load which makes it much easier to run grids of haloes which
vary in a non-trivial way (e.g. masses which increase geometrically).